Cleaning up unit tests
