+++ Tests/MathUnitTests/MathUnitTest.cs
@@ -666,86 +666,6 @@ namespace MathUnitTests
         }
 
         [TestMethod]
-        public static void Test_Max_2()
-        {
-            //
-            // Summary:
-            //     Returns the larger of two double-precision floating-point numbers.
-            //
-            // Parameters:
-            //   val1:
-            //     The first of two double-precision floating-point numbers to compare.
-            //
-            //   val2:
-            //     The second of two double-precision floating-point numbers to compare.
-            //
-            // Returns:
-            //     Parameter val1 or val2, whichever is larger. If val1 OR both val1
-            //     and val2 are equal to System.Double.NaN, System.Double.NaN is returned.
-            double[] x = new double[] { 6.00000000000000000000, 6.50000000000000000000, 7.00000000000000000000, 7.50000000000000000000, -0.50000000000000000000, -1.00000000000000000000, -1.50000000000000000000, -2.00000000000000000000 };
-            double[] y = new double[] { 1, 1.140238321, 1.600286858, 2.509178479, 4.121836054, 6.890572365, 11.59195328, -0.50000000000000000000 };
-
-            double[] answer = new double[] { 6.00000000000000000000, 6.50000000000000000000, 7.00000000000000000000, 7.50000000000000000000, 4.12183605386995000000, 6.89057236497588000000, 11.59195328, -0.50000000000000000000 };
-            double res;
-
-            for (int i = 0; i < x.Length; i++)
-            {
-                res = Math.Max(x[i], y[i]);
-
-                Assert.IsFalse((answer[i] - res) > 0.0001d || (answer[i] - res) < -0.0001d, $"Max(...{x[i]}, {y[i]}) -- FAILED AT: {res}");
-            }
-
-            res = Math.Max(10, double.NaN);
-            Assert.IsTrue(double.IsNaN(res), $"Max(...10,double.NaN) -- FAILED AT: {res}");
-
-            res = Math.Max(double.NaN, 10);
-            Assert.IsTrue(double.IsNaN(res), $"Max(...NaN, 10) -- FAILED AT: {res}");
-
-            res = Math.Max(double.NaN, double.NaN);
-            Assert.IsTrue(double.IsNaN(res), $"Max(...NaN,double.NaN) -- FAILED AT:  {res}");
-        }
-
-        [TestMethod]
-        public static void Test_Min_2()
-        {
-            //
-            // Summary:
-            //     Returns the smaller of two double-precision floating-point numbers.
-            //
-            // Parameters:
-            //   val1:
-            //     The first of two double-precision floating-point numbers to compare.
-            //
-            //   val2:
-            //     The second of two double-precision floating-point numbers to compare.
-            //
-            // Returns:
-            //     Parameter val1 or val2, whichever is smaller. If val1, val2, or both val1
-            //     and val2 are equal to System.Double.NaN, System.Double.NaN is returned.
-            double[] x = new double[] { 6.00000000000000000000, 6.50000000000000000000, 7.00000000000000000000, 7.50000000000000000000, -0.50000000000000000000, -1.00000000000000000000, -1.50000000000000000000, -2.00000000000000000000 };
-            double[] y = new double[] { 1, 1.140238321, 1.600286858, 2.509178479, 4.121836054, 6.890572365, 11.59195328, -0.50000000000000000000 };
-
-            double[] answer = new double[] { 1, 1.140238321, 1.600286858, 2.509178479, -0.50000000000000000000, -1.00000000000000000000, -1.50000000000000000000, -2.00000000000000000000 };
-            double res;
-
-            for (int i = 0; i < x.Length; i++)
-            {
-                res = Math.Min(x[i], y[i]);
-
-                Assert.IsFalse((answer[i] - res) > 0.0001d || (answer[i] - res) < -0.0001d, $"Min(...{x[i]}, {y[i]}) -- FAILED AT: {res}");
-            }
-
-            res = Math.Min(10, double.NaN);
-            Assert.IsTrue(double.IsNaN(res), $"Min(...10,double.NaN) -- FAILED AT:  {res}");
-
-            res = Math.Min(double.NaN, 10);
-            Assert.AreEqual(double.IsNaN(res), 10, $"Min(...NaN, 10) -- FAILED AT:  {res}");
-
-            res = Math.Min(double.NaN, double.NaN);
-            Assert.IsTrue(double.IsNaN(res), $"Min(...NaN,double.NaN) -- FAILED AT:  {res}");
-        }
-
-        [TestMethod]
         public static void Test_Pow_2()
         {
             //
