--- Tests/MathUnitTests/Math_Max_Tests.cs
@@ -1,74 +1,0 @@
-﻿using System;
-using nanoFramework.TestFramework;
-
-namespace MathUnitTests
-{
-    [TestClass]
-    public class Math_Max_Tests
-    {
-        [TestMethod]
-        public void Max_Double_returns_greater_value()
-        {
-            // TODO: Maybe I'm missing something here but the change in precision here seems like a defect as my expectation is that the same value sent in is returned
-            Assert.SkipTest("Skipping pending further investigation");
-
-            var expect = 1234.5678d;
-            var lower = expect / 2;
-
-            Assert.AreEqual(expect, Math.Max(expect, lower));
-            Assert.AreEqual(expect, Math.Max(lower, expect));
-        }
-
-        [TestMethod]
-        public void Max_Double_returns_NaN_if_val1_is_NaN()
-        {
-            var actual = Math.Max(double.NaN, Math.PI);
-
-            Assert.IsTrue(double.IsNaN(actual));
-        }
-
-        [TestMethod]
-        public void Max_Double_returns_NaN_if_val2_is_NaN()
-        {
-            var actual = Math.Max(Math.PI, double.NaN);
-
-            Assert.IsTrue(double.IsNaN(actual));
-        }
-
-        [TestMethod]
-        public void Max_Float_returns_greater_value()
-        {
-            var expect = 1234.5678f;
-            var lower = expect / 2;
-
-            Assert.AreEqual(expect, Math.Max(expect, lower));
-            Assert.AreEqual(expect, Math.Max(lower, expect));
-        }
-
-        [TestMethod]
-        public void Max_Float_returns_NaN_if_val1_is_NaN()
-        {
-            var actual = Math.Max(float.NaN, (float)Math.PI);
-
-            Assert.IsTrue(float.IsNaN(actual));
-        }
-
-        [TestMethod]
-        public void Max_Float_returns_NaN_if_val2_is_NaN()
-        {
-            var actual = Math.Max((float)Math.PI, float.NaN);
-
-            Assert.IsTrue(float.IsNaN(actual));
-        }
-
-        [TestMethod]
-        public void Max_Int_returns_greater_value()
-        {
-            var expect = 12345678;
-            var lower = expect / 2;
-
-            Assert.AreEqual(expect, Math.Max(expect, lower));
-            Assert.AreEqual(expect, Math.Max(lower, expect));
-        }
-    }
-}
